Refactor db creation
